Extraction of networking logic on QuesitonDetails screen
